extract analytics functions to file
